requirements.txt
flask
waitress
numpy
matplotlib
python-dotenv